requirements.txt
blinker==1.8.2
cachelib==0.9.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
filelock==3.13.1
flask==3.0.3
flask-cors==3.0.10
flask-session==0.8.0
flask-sqlalchemy==3.1.1
gmpy2==2.1.2
greenlet==3.0.1
idna==3.7
itsdangerous==2.2.0
jinja2==3.1.4
msgspec==0.18.6
networkx==3.3
numpy==1.26.4
pillow==10.4.0
pip==24.2
psycopg2-binary==2.9.9
pysocks==1.7.1
python-dotenv==1.0.1
pytorch==2.4.1
pytorch-cuda==11.8
pyyaml==6.0.1
requests==2.32.3
setuptools==72.1.0
six==1.16.0
sqlalchemy==2.0.30
sympy==1.13.2
torchaudio==2.4.1
torchtriton==3.0.0
torchvision==0.19.1
typing-extensions==4.12.2
typing_extensions==4.12.2
urllib3==2.2.2
werkzeug==3.0.3
wheel==0.44.0
